+++ addOns/callhome/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Maintenance changes.
+- Add client stats to telemetry.
 
 ## [0.7.0] - 2023-07-11
 ### Changed