+++ addOns/callhome/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update minimum ZAP version to 2.14.0.
 - Maintenance changes.
 - Add client stats to telemetry.
+- Add config stats to telemetry.
 
 ## [0.7.0] - 2023-07-11
 ### Changed